seeagent/planning-poker-platform/agent-server/requirements.txt
a2wsgi==1.10.8
absl-py==2.1.0
accelerate==1.3.0
aider-install==0.1.3
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
bentoml==1.4.7
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
cattrs==23.1.2
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-option-group==0.5.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.12
cut-cross-entropy==25.1.1
cycler==0.12.1
DataProperty==1.1.0
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.18
deprecation==2.1.0
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
et_xmlfile==2.0.0
evaluate==0.4.3
execnet==2.1.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.17.0
firecrawl==1.12.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.2.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
gotrue==2.12.0
groq==0.13.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hf_transfer==0.1.9
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-ws==0.7.2
huggingface-hub==0.28.1
hyperframe==6.1.0
identify==2.6.7
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kantoku==0.18.3
kiwisolver==1.4.8
-e git+https://github.com/EleutherAI/lm-evaluation-harness@52df63b7b30da53c481ed9090598d9189fab1d91#egg=lm_eval
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mbstrdecoder==1.1.4
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy==1.15.0
mypy-extensions==1.0.0
narwhals==1.27.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==2.2.0
nvidia-ml-py==12.570.86
oauthlib==3.2.2
ollama==0.4.7
openai==1.59.3
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-aiohttp-client==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
pip-requirements-parser==32.0.1
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
portalocker==3.1.1
postgrest==1.0.1
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.3.1
proto-plus==1.26.0
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pytablewriter==1.2.1
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.1.0
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.5.1
safetensors==0.5.2
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.8.0
shtab==1.7.1
simple-di==0.1.5
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.1
supafunc==0.9.4
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli_w==1.2.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
tqdm-multiprocess==0.0.11
traitlets==5.14.3
transformers==4.48.3
trl==0.14.0
typeguard==4.4.1
typepy==1.3.4
typing_extensions==4.12.2
tyro==0.9.13
tzdata==2024.2
unsloth_zoo==2025.2.3
uritemplate==4.1.1
urllib3==2.2.3
uv==0.7.2
uvicorn==0.34.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
word2number==1.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
